withQueryModels: Move types to QueryModel and utils to utils.ts
